Improve UI, add logout button in profile bar, add change trader link
